+++ changelog.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased] Unreleased
 
+## [2.1.4] 2018-05-08;
+### Fixed
+- the `comment_count` of posts to which comments were added using `WPDb::haveCommentInDatabase` (thanks @ptrkcsk)
+
+### Added
+- the `WPDb::countRowsInDatabase` method
+
 ## [2.1.3] 2018-05-07;
 ### Fixed
 - check for an existing `.env` file when initializing and ask for the `.env` file name in the init command
@@ -779,7 +786,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Reference to ModuleConfigException class in WPLoader class.
 
-[unreleased]: https://github.com/lucatume/wp-browser/compare/2.1.3...HEAD
+[unreleased]: https://github.com/lucatume/wp-browser/compare/2.1.4...HEAD
+[2.1.4]: https://github.com/lucatume/wp-browser/compare/2.1.3...2.1.4
 [2.1.3]: https://github.com/lucatume/wp-browser/compare/2.1.2...2.1.3
 [2.1.2]: https://github.com/lucatume/wp-browser/compare/2.1.1...2.1.2
 [2.1.1]: https://github.com/lucatume/wp-browser/compare/2.1...2.1.1